Move measure state configuration into measurePage
