+++ app/components/appMainModule/measurePageModule/index.js
@@ -5,10 +5,12 @@ require('../../common/measureCommitteeOpposing');
 require('../../common/measureCommitteeSupporting');
 
 module.exports = angular.module('measurePageModule', [
+  'ui.router',
   'measureListing',
   'measureCommitteeOpposing',
   'measureCommitteeSupporting'
 ])
   .controller('measurePageController', require('./measurePageController'))
   .controller('measureCommitteeOpposingPageController', require('./measureCommitteeOpposingPageController'))
-  .controller('measureCommitteeSupportingPageController', require('./measureCommitteeSupportingPageController'));
+  .controller('measureCommitteeSupportingPageController', require('./measureCommitteeSupportingPageController'))
+  .config(require('./state'));
